Integrate security functions into the main production script and remove the redundant engine file.
This commit refactors the codebase by moving the `verify_json_integrity` and the data shrinkage guard logic from the temporary `uretim_motoru.py` into the main `uretim_scripti.py`.

Key changes:
- `uretim_scripti.py` now performs a SHA256 integrity check on the JSON config at startup.
- The state management function, `durum_yonetimi`, has been updated with the data shrinkage guard to prevent accidental data loss.
- The now-redundant `uretim_motoru.py` has been deleted.
- The test script `test_guvenlik.py` has been updated to import from and test the consolidated `uretim_scripti.py`.

All dependencies have been installed and the test suite now passes, confirming the successful integration.
